move k8 apis to k8 package
